# Aula 07/doc/openapi.yaml
...
openapi: 3.0.3
info:
  title: Projeto de Filmes - SENAI Jandira 
  description: API destinada para controlar filmes
  termsOfService: http://swagger.io/terms/
  version: 1.0.0
...
servers:
  - url: http://
paths:
...
  /v1/senai/locadora/filme:
    post:
      tags:
        - Filme
      summary: Adiciona um novo filme
      description: Adiciona um novo filme no Banco de Dados
      requestBody:
        description: Registra um filme
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme'
        required: true
      responses:
        '201':
          description: Registro inserido com sucesso
        '415':
          description: Não foi possível processar a requisição, pois o formato de dados aceito pela API é somente JSON
        '500':
          description: Erro na Model ou Controller

    get:
      tags:
        - Filme
      summary: Lista todos os filmes
      description: Listagem de todos os filmes registrados no sistema
      responses:
        '200':
          description: Busca feita com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Filme'          
           
        '404':
          description: Não foi possível encontrar nenhum dado para retorno
        '500':
          description: Erro na Model ou Controller
#---------------------------------------------------------------#

#----------[ ver, atualizar e deletar filme pelo ID ]-----------#
  /v1/senai/locadora/filme/{id}:
    get:
      tags:
        - Filme
      summary: Busca um filme pelo seu ID
      description: Retorna um filme pelo ID
      parameters:
        - name: id
          in: path
          description: Retorna um filme pelo seu ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Busca feita com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Filme'          
        '400':
          description: Os dados enviados na requisição não estão corretos
        '404':
          description: Não foi possível encontrar nenhum dado para retorno
        '500':
          description: Erro na Model ou Controller

    put:
      tags:
        - Filme
      summary: Atualizar um filme existente pelo ID
      description: Localizar e atualizar um filme buscando pelo ID
      parameters:
        - name: id
          in: path
          description: ID para localizar e atualizar o filme
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Dados para atualizar o filme
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme'
        required: true
      responses:
        '200':
          description: Registro atualizado com sucesso
        '400':
          description: Os dados enviados na requisição não estão corretos
        '404':
          description: Não foi possível encontrar nenhum dado para retorno
        '415':
          description: Não foi possível processar a requisição, pois o formato de dados aceito pela API é somente JSON
        '500':
          description: Erro na Model ou Controller
       
    delete:
      tags:
        - Filme
      summary: Excluir um filme existente pelo ID
      description: Excluir um filme pelo ID
      parameters:
        - name: id
          in: path
          description: ID para localizar e excluir o filme
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Registro excluído com sucesso
        '400':
          description: Os dados enviados na requisição não estão corretos
        '404':
          description: Não foi possível encontrar nenhum dado para retorno
        '500':
          description: Erro na Model ou Controller
#---------------------------------------------------------------#
 
#-------[ inserir e listar classificação indicativa ]-----------#
  /v1/senai/locadora/indicativa_classificacao:
    post:
      tags:
        - ClassificacaoIndicativa
      summary: Adiciona uma nova classificação indicativa
      description: Adiciona uma nova classificação indicativa no Banco de Dados
      requestBody:
        description: Registra uma classificação indicativa
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassificacaoIndicativa'
        required: true
      responses:
        '201':
          description: Registro inserido com sucesso
        '415':
          description: Não foi possível processar a requisição, pois o formato de dados aceito pela API é somente JSON
        '500':
          description: Erro na Model ou Controller

    get:
      tags:
        - ClassificacaoIndicativa
      summary: Lista todas as classificações indicativas
      description: Listagem de todas as classificações indicativas registradas no sistema
      responses:
        '200':
          description: Busca feita com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClassificacaoIndicativa'          
           
        '404':
          description: Não foi possível encontrar nenhum dado para retorno
        '500':
          description: Erro na Model ou Controller
#---------------------------------------------------------------#  

#-[ ver, atualizar e deletar classificação indicativa pelo ID ]-#
  /v1/senai/locadora/indicativa_classificacao/{id}:
    get:
      tags:
        - ClassificacaoIndicativa
      summary: Busca uma classificação indicativa pelo seu ID
      description: Retorna uma classificação indicativa pelo ID
      parameters:
        - name: id
          in: path
          description: Retorna uma classificação indicativa pelo seu ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Busca feita com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassificacaoIndicativa'          
        '400':
          description: Os dados enviados na requisição não estão corretos
        '404':
          description: Não foi possível encontrar nenhum dado para retorno
        '500':
          description: Erro na Model ou Controller

    put:
      tags:
        - ClassificacaoIndicativa
      summary: Atualizar uma classificação indicativa existente pelo ID
      description: Localizar e atualizar uma classificação indicativa buscando pelo ID
      parameters:
        - name: id
          in: path
          description: ID para localizar e atualizar a classificação indicativa
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Dados para atualizar a classificação indicativa
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassificacaoIndicativa'
        required: true
      responses:
        '200':
          description: Registro atualizado com sucesso
        '400':
          description: Os dados enviados na requisição não estão corretos
        '404':
          description: Não foi possível encontrar nenhum dado para retorno
        '415':
          description: Não foi possível processar a requisição, pois o formato de dados aceito pela API é somente JSON
        '500':
          description: Erro na Model ou Controller
       
    delete:
      tags:
        - ClassificacaoIndicativa
      summary: Excluir uma classificação indicativa existente pelo ID
      description: Excluir uma classificação indicativa pelo ID
      parameters:
        - name: id
          in: path
          description: ID para localizar e excluir a classificação indicativa
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Registro excluído com sucesso
        '400':
          description: Os dados enviados na requisição não estão corretos
        '404':
          description: Não foi possível encontrar nenhum dado para retorno
        '500':
          description: Erro na Model ou Controller
#---------------------------------------------------------------#

#------------------[ inserir e listar gênero ]------------------#
  /v1/senai/locadora/genero:
    post:
      tags:
        - Genero
      summary: Adiciona um novo gênero de filme
      description: Adiciona um novo gênero de filme no Banco de Dados
      requestBody:
        description: Registra um gênero de filme
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Genero'
        required: true
      responses:
        '201':
          description: Registro inserido com sucesso
        '415':
          description: Não foi possível processar a requisição, pois o formato de dados aceito pela API é somente JSON
        '500':
          description: Erro na Model ou Controller

    get:
      tags:
        - Genero
      summary: Lista todos os gêneros do filme
      description: Listagem de todos os gêneros do filme registrados no sistema
      responses:
        '200':
          description: Busca feita com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Genero'          
           
        '404':
          description: Não foi possível encontrar nenhum dado para retorno
        '500':
          description: Erro na Model ou Controller
#---------------------------------------------------------------#

#----------[ ver, atualizar e deletar gênero pelo ID ]----------#
  /v1/senai/locadora/genero/{id}:
    get:
      tags:
        - Genero
      summary: Busca um gênero pelo seu ID
      description: Retorna um gênero pelo ID
      parameters:
        - name: id
          in: path
          description: Retorna um gênero pelo seu ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Busca feita com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Genero'          
        '400':
          description: Os dados enviados na requisição não estão corretos
        '404':
          description: Não foi possível encontrar nenhum dado para retorno
        '500':
          description: Erro na Model ou Controller

    put:
      tags:
        - Genero
      summary: Atualizar um gênero existente pelo ID
      description: Localizar e atualizar um gênero buscando pelo ID
      parameters:
        - name: id
          in: path
          description: ID para localizar e atualizar o gênero
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Dados para atualizar o gênero
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Genero'
        required: true
      responses:
        '200':
          description: Registro atualizado com sucesso
        '400':
          description: Os dados enviados na requisição não estão corretos
        '404':
          description: Não foi possível encontrar nenhum dado para retorno
        '415':
          description: Não foi possível processar a requisição, pois o formato de dados aceito pela API é somente JSON
        '500':
          description: Erro na Model ou Controller
       
    delete:
      tags:
        - Genero
      summary: Excluir um gênero existente pelo ID
      description: Excluir um gênero pelo ID
      parameters:
        - name: id
          in: path
          description: ID para localizar e excluir o gênero
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Registro excluído com sucesso
        '400':
          description: Os dados enviados na requisição não estão corretos
        '404':
          description: Não foi possível encontrar nenhum dado para retorno
        '500':
          description: Erro na Model ou Controller       
#---------------------------------------------------------------#

#------------------[ inserir e listar elenco ]------------------#
...
components:
  schemas:
    Filme:
      required:
        - nome
        - data_lancamento
        - duracao
        - sinopse
        - valor

      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        data_lancamento:
          type: string
          format: date
        duracao:
          type: string
          format: time
        sinopse:
          type: string
        avaliacao:
          type: number
        valor:
          type: number
        capa:
          type: string
    ClassificacaoIndicativa:
      required:
        - codigo
        - nome
        - descricao

      type: object
      properties:
        id:
          type: integer
        codigo:
          type: string
        nome:
          type: string
        descricao:
          type: string
    Genero:
      required:
        - nome

      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
...